Added support for appending objects to log messages
Objects can be appended to log messages as part of the supplant array:

logger.info('this object caused at line {0} caused trouble:', ['18', {'foo': 'bar'}]);

All objects not being referenced in the supplant array by the curly brackets are being logged at the end of the message. The original functionality is preserved.